SLEMA-188 Moved LinkButton and MenuOptionDivider from an invalid directory
