+++ SlemaForAndroid/lib/features/settings/presentation/screen/settings_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:pg_slema/features/settings/presentation/screen/developer_settings_screen.dart';
-import 'package:pg_slema/features/well_being/presentation/widget/menu_option_divider.dart';
+import 'package:pg_slema/utils/widgets/menu_option_divider.dart';
 import 'package:pg_slema/features/well_being/presentation/widget/navigator_push_link_button.dart';
 import 'package:pg_slema/utils/widgets/default_appbar/default_appbar.dart';
 import 'package:pg_slema/utils/widgets/default_body/default_body.dart';
