+++ SlemaForAndroid/lib/utils/widgets/menu_option_divider.dart
